# (#3557) - idb allDocs() refactor, perf improvements