--- a/lib/adapters/idb/idb-bulk-docs.js
+++ b/lib/adapters/idb/idb-bulk-docs.js
@@ -26,6 +26,7 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
   var attachStore;
   var attachAndSeqStore;
   var docInfoError;
+  var docCountDelta = 0;
 
   for (var i = 0, len = docInfos.length; i < len; i++) {
     var doc = docInfos[i];
@@ -45,7 +46,7 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
   var results = new Array(docInfos.length);
   var fetchedDocs = new utils.Map();
   var preconditionErrored = false;
-  var blobType = api._blobSupport ? 'blob' : 'base64';
+  var blobType = api._meta.blobSupport ? 'blob' : 'base64';
 
   utils.preprocessAttachments(docInfos, blobType, function (err) {
     if (err) {
@@ -128,8 +129,8 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
       return;
     }
 
-    Changes.notify(api._name);
-    api._docCount = -1; // invalidate
+    Changes.notify(api._meta.name);
+    api._meta.docCount += docCountDelta;
     callback(null, results);
   }
 
@@ -186,6 +187,8 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
 
   function writeDoc(docInfo, winningRev, deleted, callback, isUpdate,
                     delta, resultsIdx) {
+
+    docCountDelta += delta;
 
     var doc = docInfo.data;
     doc._id = docInfo.metadata.id;