- keep docCount in memory, ala leveldb.js
- use metadata.deleted instead of calculating
  with utils.isDeleted()
- fetch docs asynchronously for include_docs=true