@nolanlawson 2015-02-21
Currently this is showing a big improvement for Chrome and an almost equal un-improvement for Firefox :(

```
temp-views test, 10 iterations
Chrome 42: 89105ms -> 78397ms
Firefox 37:  80500ms -> 86922ms
Firefox 37 (test 2): 82754ms -> 89679ms
```

I've been trying to figure out what's going on, because it doesn't make any sense to me that it would be so much slower in one but faster in the other.