@nolanlawson 2015-02-21
Some more numbers from Chrome:

```
78558ms -> 83279ms
79023ms -> 78912ms
```

I think what this is revealing is just that our performance tests aren't fine-grained or consistent enough to tell us anything useful. For this PR, I'm happy to go with my gut.