@nolanlawson 2015-02-21
OK, I've added a new performance test, which I think more accurately tests what I'm trying to improve here: `all-docs-include-docs`:

```
Chrome: 11256ms -> 6982ms
Firefox: 7388ms -> 5645ms
```