@daleharvey 2015-02-25
Slightly wary about the caching, but with tests passing and the performance improvements I think this is good, cheers - https://github.com/pouchdb/pouchdb/commit/f010d2eb959e045d9c94462780207f1ea16e45c0