@nolanlawson 2015-02-25
The caching of `docCount` is the exact same trick used by `leveldb.js`, so it has precedent. I'd also like to add it to WebSQL eventually, but these days I'm more concerned about IDB improvement.